Restructure context
