+++ packages/playground/src/app/Editor/index.tsx
@@ -1,8 +1,8 @@
 import { useState, useContext, ReactElement } from 'react';
 import MonacoEditor, { MonacoEditorProps } from 'react-monaco-editor';
 import exampleCode from './example-code';
-import { AppContext } from '~/context';
-import { Dispatch, State } from '~/context/reducer';
+import { AppContext } from '~/context/app/';
+import { Dispatch, State } from '~/context/app/reducer';
 
 export const Editor = (): ReactElement => {
   const [code, setCode] = useState(exampleCode);
