+++ packages/playground/src/app/Header/SettingsSubmenu.tsx
@@ -1,7 +1,7 @@
 import { LabeledInputSwitch } from '@paritytech/components/';
 import { ReactElement, useContext } from 'react';
-import { AppContext } from '~/context';
-import { Dispatch, State } from '~/context/reducer';
+import { AppContext } from '~/context/app/';
+import { Dispatch, State } from '~/context/app/reducer';
 
 export const SettingsSubmenu = (): ReactElement => {
   const [state, dispatch]: [State, Dispatch] = useContext(AppContext);
