+++ packages/playground/src/app/index.tsx
@@ -2,13 +2,16 @@ import { Console } from './Console';
 import { Editor } from './Editor';
 import { Layout } from './Layout';
 import { Header } from './Header';
-import { AppProvider } from '~/context';
+import { AppProvider } from '~/context/app/';
+import { MessageProvider } from '../context/messages/';
 import { ReactElement } from 'react';
 
 const App = (): ReactElement => {
   return (
     <AppProvider>
-      <Layout Header={Header} Editor={Editor} Console={Console} />
+      <MessageProvider>
+        <Layout Header={Header} Editor={Editor} Console={Console} />
+      </MessageProvider>
     </AppProvider>
   );
 };
